projects/04-urban-mobility-analytics/requirements.txt
pandas>=2.2
pyarrow>=16.0
duckdb>=1.0
prefect>=2.14
requests>=2.31
python-dotenv>=1.0
pytest>=8.0
